--- a/benchmarks/go.mod
+++ b/benchmarks/go.mod
@@ -6,6 +6,6 @@ replace github.com/goccy/go-yaml => ../
 
 require (
 	github.com/goccy/go-yaml v0.0.0-00010101000000-000000000000
-	gopkg.in/yaml.v2 v2.3.0
-	gopkg.in/yaml.v3 v3.0.0-20200506231410-2ff61e1afc86
+	gopkg.in/yaml.v2 v2.4.0
+	gopkg.in/yaml.v3 v3.0.1
 )