Improve parser performance ( inspired by #438, Thanks )

resolve https://github.com/goccy/go-yaml/issues/325

30% Speed up

### BEFORE

```
goos: darwin
goarch: arm64
pkg: benchmarks
Benchmark/gopkg.in/yaml.v2-10             244044              4919 ns/op
Benchmark/gopkg.in/yaml.v3-10             167894              7052 ns/op
Benchmark/github.com/goccy/go-yaml-10      93895             12559 ns/op
```

### AFTER

```
goos: darwin
goarch: arm64
pkg: benchmarks
Benchmark/gopkg.in/yaml.v2-10             242497              4945 ns/op
Benchmark/gopkg.in/yaml.v3-10             167689              6988 ns/op
Benchmark/github.com/goccy/go-yaml-10     126858              9333 ns/op
```